HW7 with alpha anim
