+++ app/src/main/java/com/dimsonpip/loftmoney/screens/main/MainActivity.java
@@ -51,7 +51,6 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onResume() {
-        overridePendingTransition(android.R.anim.slide_in_left, android.R.anim.slide_out_right);
         super.onResume();
     }
 }
