+++ app/src/main/res/values/styles.xml
@@ -18,10 +18,4 @@
         <item name="android:textColor">@color/blue_background_text_color</item>
     </style>
 
-    <style name="ItemStyle" parent="TextAppearance.AppCompat">
-        <item name="fontFamily">sans-serif-medium</item>
-        <item name="android:textSize">16sp</item>
-        <item name="android:textColor">#666664</item>
-    </style>
-
 </resources>
